Solved header include order problem
The problem arose because several headers had structs that were dependant on each other. Depending on the order the files where compiled/linked some of these structs would have no definition at the point when a header that relied on that struct was linked. This resulted in a number of errors. I solved this by moving all struct definitions to one header file that is then included in all files that use one or more of these structs.